backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
motor
bcrypt
python-jose[cryptography]==3.3.0
python-multipart
pydantic[email]
